pineapple/requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
brotlicffi==1.1.0.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
mysqlclient==2.2.7
packaging==25.0
pip==25.1
priority==2.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==25.1.0
PySocks==1.7.1
redis==6.2.0
requests==2.32.4
service-identity==24.2.0
setuptools==78.1.1
shortuuid==1.0.13
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uWSGI==2.0.30
wheel==0.45.1
wsproto==1.2.0
zope.interface==7.2
